refactor(web): Moved Avatar Logo menu to a child component
